remove warnings

+++ src/rws_frame.h
@@ -48,7 +48,7 @@ typedef struct _rws_frame_struct
 	unsigned char header_size;
 } _rws_frame;
 
-unsigned int rws_check_recv_frame_size(const void *data,const size_t data_size);
+size_t rws_check_recv_frame_size(const void *data,const size_t data_size);
 
 _rws_frame * rws_frame_create_with_recv_data(const void * data, const size_t data_size);
